+++ spec/features/staff/users_can_create_a_comment_spec.rb
@@ -21,7 +21,7 @@ RSpec.describe "Users can create a comment" do
         context "when the report is editable" do
           scenario "the user cannot add a comment" do
             visit report_path(report)
-            click_on t("tabs.report.variance")
+            click_on t("tabs.report.variance.heading")
             expect(page).not_to have_content t("table.body.report.add_comment")
           end
         end
@@ -30,7 +30,7 @@ RSpec.describe "Users can create a comment" do
           let(:report) { create(:report, fund: activity.associated_fund, organisation: delivery_partner_user.organisation) }
           scenario "the user cannot add a comment" do
             visit report_path(report)
-            click_on t("tabs.report.variance")
+            click_on t("tabs.report.variance.heading")
             expect(page).not_to have_content t("table.body.report.add_comment")
           end
         end
@@ -65,7 +65,7 @@ RSpec.describe "Users can create a comment" do
           let(:report) { create(:report, :approved, fund: activity.associated_fund, organisation: delivery_partner_user.organisation) }
           scenario "the user cannot add a comment" do
             visit report_path(report)
-            click_on t("tabs.report.variance")
+            click_on t("tabs.report.variance.heading")
             expect(page).not_to have_content t("table.body.report.add_comment")
           end
         end